refactor erc1155 fixtures
