+++ tests/test_packages/test_connections/test_ledger/test_ledger_api.py
@@ -215,8 +215,7 @@ async def test_send_signed_transaction_ethereum(ledger_apis_connection: Connecti
     assert type(response.message) == LedgerApiMessage
     response_message = cast(LedgerApiMessage, response.message)
     assert (
-        response_message.performative
-        != LedgerApiMessage.Performative.ERROR
+        response_message.performative != LedgerApiMessage.Performative.ERROR
     ), f"Received error: {response_message.message}"
     assert (
         response_message.performative
