+++ tests/test_packages/test_contracts/test_erc1155/test_contract.py
@@ -18,39 +18,15 @@
 # ------------------------------------------------------------------------------
 
 """The tests module contains the tests of the packages/contracts/erc1155 dir."""
-
-import json
-from pathlib import Path
-from typing import cast
-
-# from unittest import mock
-
 import pytest
 
-
-from aea.configurations.base import (
-    ComponentConfiguration,
-    ComponentType,
-    ContractConfig,
-)
-from aea.contracts import contract_registry
-from aea.contracts.base import Contract
 from aea.crypto.ethereum import EthereumCrypto
 from aea.crypto.registries import crypto_registry, ledger_apis_registry
-
-# from aea.helpers.transaction.base import (
-#     RawTransaction,
-#     SignedTransaction,
-#     Terms,
-#     TransactionDigest,
-#     TransactionReceipt,
-# )
 
 from tests.conftest import (
     ETHEREUM_ADDRESS_ONE,
     ETHEREUM_ADDRESS_TWO,
     ETHEREUM_TESTNET_CONFIG,
-    ROOT_DIR,
 )
 
 ledger = [
@@ -74,36 +50,6 @@ def crypto_api(request):
     crypto_id = request.param[0]
     api = crypto_registry.make(crypto_id)
     yield api
-
-
-@pytest.fixture()
-def erc1155_contract():
-    directory = Path(ROOT_DIR, "packages", "fetchai", "contracts", "erc1155")
-    configuration = ComponentConfiguration.load(ComponentType.CONTRACT, directory)
-    configuration._directory = directory
-    configuration = cast(ContractConfig, configuration)
-
-    # TODO some other tests don't deregister contracts from the registry.
-    #   find a neater solution.
-    if configuration.public_id in contract_registry.specs.keys():
-        contract_registry.specs.pop(str(configuration.public_id))
-
-    # load contract into sys modules!
-    Contract.from_config(configuration)
-
-    path = Path(configuration.directory, configuration.path_to_contract_interface)
-    with open(path, "r") as interface_file:
-        contract_interface = json.load(interface_file)
-
-    contract_registry.register(
-        id_=str(configuration.public_id),
-        entry_point=f"{configuration.prefix_import_path}.contract:{configuration.class_name}",
-        class_kwargs={"contract_interface": contract_interface},
-        contract_config=configuration,
-    )
-    contract = contract_registry.make(configuration.public_id)
-    yield contract
-    contract_registry.specs.pop(str(configuration.public_id))
 
 
 @pytest.mark.network
